+++ scripts/build.sh
@@ -1,0 +1,3 @@
+#!/bin/bash
+
+docker build -t wtfcoderz/bigdata4all . && docker-compose up -d
